requirements.txt
requests>=2.28.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
